refactor: :recycle: Update the Id type to int in base entity

+++ Core/Entities/Block/BlockChain.cs
@@ -23,7 +23,7 @@ public class BlockChain : BaseEntity
     public NotificationType NotiTypes { get; set; }
 
     /* Foreign Key for Auditor */
-    public long IdAuditorFk { get; set; }
+    public int IdAuditorFk { get; set; }
     public Auditor Auditors { get; set; }
 
 }
